Use opts.quiet when determining return value
This makes it possible for Grunt to return a non-0 exit code when lint warnings are present.

If `opts.quiet` is set to `false` (default), then this task will return `false` if there are any errors *or* warnings. If `opts.quiet` is `true`, then this task will only return `false` if there are errors.