Add a force Option to allow the build to continue
In my use case for grunt scripts we need to have the ability to generate all the errors for our different modules separately. Currently in the existing framework if one module has one error it stops the whole build and we get zero feedback and visibility into the errors for other modules.